Refatoraçao concluida.
